📝 Extensive changes to database
- Fix group members having wrong columns
- Change ids to long/int from ulong to avoid issues with conversions

+++ Komi.Data/Models/Moderation/ModerationData.cs
@@ -5,7 +5,7 @@ namespace Komi.Data.Models.Moderation
 {
     public class ModerationData
     {
-        public ulong ModerationDataId { get; set; }
+        public int ModerationDataId { get; set; }
 
         public List<WarningData>? Warnings { get; set; } = new List<WarningData>();
